fix(src/Plazza/ Restaurant|Reception): little changes made in merge where incorrect so fix it

+++ src/Plazza/Product/Pizza/PizzaType/Regina/Regina.hpp
@@ -14,7 +14,7 @@ namespace Pizzeria
 {
     class ReginaPizza : public Pizza {
       public:
-        ReginaPizza(PizzaSize size, long multiplier);
+        ReginaPizza(PizzaSize size, double multiplier);
         ~ReginaPizza() override = default;
     };
 } // namespace Pizzeria
